requirements.txt
Flask==2.3.3
flask-cors==4.0.0
supabase==1.2.0
httpx==0.24.1
python-dotenv==1.0.0
waitress==2.1.2
aiohttp==3.9.1
asyncio
mcp
